requirements.txt
2to3==1.0
absl-py==0.9.0
alabaster==0.7.12
altair==4.1.0
amqp==2.5.2
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appdirs==1.4.4
argh==0.26.2
asgiref==3.2.10
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.3
Babel==2.7.0
backcall==0.2.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.0.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.5.0.5
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.5
blinker==1.4
bokeh==1.3.4
boto==2.49.0
boto3==1.14.7
botocore==1.17.7
Bottleneck==1.2.1
bs4==0.0.1
cachetools==4.1.0
caesarcipher==1.0
celery==4.2.1
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.8.3
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
distributed==2.5.2
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap-form==3.4
django-environ==0.4.5
django-heroku==0.3.1
django-urls==1.1.2
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
ebcli==4.0.18
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.6
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.1.0
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
fsspec==0.5.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
glob2==0.7
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.27.2
gunicorn==19.9.0
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
hypothesis==5.16.1
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==1.6.1
imutils==0.5.3
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
keyrings.alt==3.4.0
kiwisolver==1.2.0
kombu==4.6.7
lazy-object-proxy==1.4.2
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.5.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
MechanicalSoup==0.12.0
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==8.4.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.3
nltk==3.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.18.5
numpydoc==0.9.1
olefile==0.46
opencv-python==4.3.0.36
openpyxl==3.0.0
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.1
parse==1.15.0
parso==0.7.0
partd==1.0.0
path==14.0.1
path.py==12.0.1
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pdfminer==20191125
pdfminer.six==20200124
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
plotly==4.5.3
pluggy==0.13.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.6.3
psycopg2==2.8.5
psycopg2-binary==2.7.7
PuLP==2.1
py==1.8.0
pybase64==1.0.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.7
pycurl==7.43.0.3
pydeck==0.4.0b2
pydocstyle==5.0.2
pyee==7.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.2
PyNaCl==1.3.0
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-constraint==1.4.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2019.3
PyWavelets==1.0.3
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
regex==2020.6.8
requests==2.22.0
requests-html==0.10.0
retrying==1.3.3
rope==0.14.0
Rtree==0.9.4
ruamel-yaml==0.15.46
s3transfer==0.3.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.5.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
sqlparse==0.3.0
statsmodels==0.10.1
streamlit==0.62.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
tflearn==0.3.2
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
tsp==0.0.7
twitter-scraper==0.4.2
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.9
validators==0.15.0
vine==1.3.0
virtualenv==20.0.7
w3lib==1.22.0
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.7.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
yapf==0.29.0
zict==1.0.0
zipp==3.1.0
